Make it explicit the optimization condition for pattern generation
The code depends upon `patterns_from(pattern, **empty_hash)` that passes
an empty hash as the second argument.  Since 2.7, it passes only one
positional argument.

I think that the intention here is to skip pattern generation when
`options` are empty and when `pattern` is a Pattern object.
This changeset makes it more explicit.