Use braces to pass a Hash object instead of keywords
The code expects `foo("a" => 1, :b => 2)` to pass one positional
argument `{"a"=>1}` and one keyword argument `{:b=>2}`, but
unfortunately, this behavior of Hsah separation was a bug of Ruby 2.6 or
prior.  Since Ruby 2.7, non-Symbol key is allowed as keyword arguments,
so both are passed as keyword arguments.

This changeset adds braces to some method calls to pass a Hash object.